Remove V2, fix phpcs
