+++ src/Form/SettingsForm.php
@@ -150,12 +150,6 @@ class SettingsForm extends ConfigFormBase {
       // we include this here to make it skip the form-level validator.
       '#validate' => [],
     ];
-    $form['notification_settings']['nexx_api_version'] = [
-      '#type' => 'select',
-      '#title' => $this->t('NEXX API version'),
-      '#options' => [2 => 'APIv2', 3 => 'APIv3'],
-      '#default_value' => $settings->get('nexx_api_version'),
-    ];
     $form['notification_settings']['info'][] = [
       '#markup' => '<p>' . $this->t('The current value to provide in omnia domain settings for the video endpoint is:<br><strong>:endpoint</strong>',
         [
@@ -196,7 +190,6 @@ class SettingsForm extends ConfigFormBase {
       ->set('omnia_id', $values['omnia_id'])
       ->set('api_secret', $values['api_secret'])
       ->set('notification_access_key', $values['notification_access_key'])
-      ->set('nexx_api_version', $values['nexx_api_version'])
       ->save();
   }
